Changed algorithm for calculating ROI for Bitcoin (different compared to ETH) Changed hash rate to scientific notation for Bitcoin network difficulty

+++ Cryptocurrencies.ROI.Calculator/Domain/ComputingDevices/ASIC.cs
@@ -8,9 +8,10 @@ namespace Cryptocurrencies.ROI.Calculator.Domain
 {
     class ASIC : ComputingDevice
     {
-        public ASIC(decimal price, string model, string manufacturer, string version, int hashPower)
+        public ASIC(decimal price, string model, string manufacturer, string version, long hashPower)
             : base(price, model, manufacturer, version, hashPower)
         {
+
 
         }
     }
